requirements.txt
# requirements.txt - Python dependencies for Travel Assistant
langchain==0.1.0
langchain-groq==0.0.1
langchain-core==0.1.23
langchain-community==0.0.20
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.2
numpy==1.26.4
tenacity==8.5.0
streamlit==1.28.0

